requirements.txt
pandas
numpy
scikit-learn
paho-mqtt
websockets
langchain
langchain-openai
langchain-community
python-dotenv